+++ nx-plugin/src/generators/angular/generator.ts
@@ -86,13 +86,19 @@ export async function angularGenerator(
       '@onecx/portal-integration-angular': oneCXLibVersion,
       '@onecx/portal-layout-styles': oneCXLibVersion,
       '@onecx/accelerator': oneCXLibVersion,
-      "@onecx/angular-accelerator": oneCXLibVersion,
+      '@onecx/angular-accelerator': oneCXLibVersion,
       '@onecx/integration-interface': oneCXLibVersion,
       '@ngx-translate/core': '^15.0.0',
       '@ngx-translate/http-loader': '^8.0.0',
       '@angular-architects/module-federation': '^18.0.4',
       '@angular/cdk': '^18.1.0',
       'keycloak-angular': '^16.0.1',
+      '@onecx/angular-remote-components': oneCXLibVersion,
+      '@onecx/angular-webcomponents': oneCXLibVersion,
+      '@onecx/angular-auth': oneCXLibVersion,
+      '@angular/elements': '^15.2.7',
+      "@angular/animations": "^15.2.7",
+      '@webcomponents/webcomponentsjs': '^2.8.0',
     },
     {
       '@openapitools/openapi-generator-cli': '^2.5.2',
@@ -253,6 +259,9 @@ function adaptAngularPrefixConfig(tree: Tree) {
   updateJson(tree, 'project.json', (json) => {
     json.prefix = 'app';
     json.targets.test.options.coverage = true;
+    json.targets.build.options.scripts = [
+      'node_modules/@webcomponents/webcomponentsjs/webcomponents-bundle.js',
+    ];
     return json;
   });
 }